Simplify configuration framework
- configuration class in test_core.py has been removed because it was
redundant (there is already a testing config in config.py)
- configuration dictionary in config.py has been removed because it adds no
value to the config framework. Instead, the default config is now specified via
an optional keyword argument in create_app().
- fix FLASK_CONFIG having no effect because there was no code to actually load
the config from a file
- the testing config now uses an in-memory database instead of an empty
URL (maybe this is actually the same thing in the end, but at least it's more
explicit now)